refactor: reorganize into gc/ with the 4 steps as files; rewrite crate/module docs
The module tree now tells the protocol's story. comp_gc -> gc; it_gc -> gc/body;
and the 1191-line extract.rs (which held step 1, step 2, AND their shared
machinery) splits along its real seams — chunk and extract never call each
other, they both call the shared "two moves":

  gc/onehot.rs  — the shared one-hot doubling tree + width-l casts + pin +
                  bare-lane types (Z2/Wide), pub(crate) to the two steps
  gc/chunk.rs   — step 1: input bits -> ring word (bin_to_word)
  gc/extract.rs — step 2: ring word -> bits + one-hot + upcast (word_to_bin_up)
  gc/fold.rs    — step 3 (moved)
  gc/body.rs    — step 4 (renamed from it_gc.rs)

lib.rs is rewritten as the canonical pipeline picture (the 4-step diagram) with
a CF/NCF glossary — the single most-cited newcomer blocker (control-friendly
Z_{2^k} labels for steps 1-3 vs non-control-friendly Z_p residues for the body),
never defined anywhere before. Dropped the "Phase 1 / Phase 3" numbering (no
Phase 2 existed) in favor of the steps-1-4 scheme used crate-wide. gc/mod.rs
gets the comp-GC vs IT-GC framing.

Pure code motion: 91 tests pass (release, debug, count-hashes); reference-
workload instruction count unchanged (16.347G vs 16.344G, within noise).
