+++ src/gc/extract.rs
@@ -1,213 +1,17 @@
-//! Chunk conversion and sub-chunk extraction: straight-line garble/eval over
-//! bare labels (the [`super::fold`] pattern, extended to width-`l` wires).
+//! Step 2 / 4 (computational GC): decompose a ring word into its sub-chunk
+//! bits + a binary one-hot + the peel upcast — the fused `word_to_bin_up`.
 //!
-//! Both steps are built from the same two moves:
-//!
-//! * **The one-hot doubling tree.** Level `m` holds the binary one-hot of the
-//!   low `m` bits. The garbler hashes EVERY slot (its pass is x-blind); the
-//!   evaluator hashes only the closed slots and solves the single open one per
-//!   level through that level's 1-bit scale join:
-//!   `y_hot = L_bit ⊕ diff ⊕ (⊕_{j≠hot} y_j)`.
-//! * **Width-`l` casts + the pin join.** Every leaf is cast once to Z_{2^l}
-//!   (`A_p = H(leaf_p)`, `l` CCRH blocks). The single width-`l` pin join
-//!   `Σ_p A_p ⋈ 1` lets the evaluator solve the whole hot class by sums:
-//!   `L_root = pin_diff` (the constant-1 wire's label is 0), so any linear
-//!   functional `Σ c_p·A_p` with a constant coefficient `q` on the unsolved
-//!   hot class evaluates as `Σ_{solved} c_p·A_p + q·(L_root − Σ_{solved} A_p)`.
-//!
-//! The extract step runs the fused `word_to_bin_up` schedule per sub-chunk:
-//! residues `x mod 2^i` from per-level accumulators peel the bits LSB-first,
-//! and the level-`k` functional `Σ p·A_p` is the peel upcast. Values are never
-//! discovered — the evaluator knows `x`, so every hot index is index
-//! arithmetic. No gate graph, no worklist, no `Label` heap traffic in the
-//! inner loops: Z₂ wires are `[u64; 2]` words and width-`l` wires are
-//! `[u32; λ]` lane arrays.
-//!
-//! Nonce discipline (paper App. A, Def. 4): Z₂ tree hashes draw bulk-domain
-//! ids and width-`l` cast hashes draw solo-domain ids, each from
-//! caller-supplied fresh windows ([`chunk_nonces`],
-//! [`extract_nonces`]); the garbler and evaluator share the window and
-//! index it by slot position, so pads agree at every closed slot.
+//! Per prime: form `r_i = Σ_c coeff_c·w_c`, split it into ≤8-bit sub-chunks,
+//! and for each run the [`super::onehot`] tree + casts. The width-`l` casts do
+//! double duty — their weighted class sums both peel the bits (LSB-first) and
+//! give the upcast `Σ p·A_p` that reconstructs the sub-chunk value for the next
+//! peel. Outputs the first sub-chunk's binary one-hot + the remaining bits,
+//! which [`super::fold`] consumes.
 
-use crate::crypto::expand;
+use super::onehot::*;
 use crate::hash;
 use crate::label::{LAMBDA, Label, delta_r};
 
-/// Packed words of a CF Z₂ label (the bare-word working type).
-pub(crate) type Z2 = [u64; 2];
-
-/// One coordinate per u32 lane of a width-`l` CF label (2 ≤ l ≤ 32).
-pub(crate) type Wide = [u32; LAMBDA];
-
-// ---- bare-label helpers ----
-
-#[inline]
-fn z2_of(l: &Label) -> Z2 {
-    debug_assert_eq!(l.modulus(), 2);
-    let raw = l.raw_bits();
-    [raw[0], raw[1]]
-}
-
-#[inline]
-fn z2_label(w: Z2) -> Label {
-    Label::from_raw_bits(vec![w[0], w[1]], 2)
-}
-
-#[inline]
-fn wide_of(l: &Label) -> Wide {
-    l.lanes().try_into().expect("lanes are LAMBDA wide")
-}
-
-#[inline]
-fn wide_label(w: &Wide, modulus: u64) -> Label {
-    let mask = (modulus - 1) as u32;
-    Label::from_lanes(w.iter().map(|&l| l & mask).collect(), modulus)
-}
-
-/// `dst += src` lane-wise, mod 2^32.
-///
-/// The lane ops run UNMASKED: wrapping u32 arithmetic is a ring homomorphism
-/// onto Z_{2^l} under truncation, so the `& (2^l − 1)` happens once at the
-/// boundaries that make wire objects canonical ([`wide_label`], the pin
-/// emission, and the Z₂ packing of bits below the preserved width — the peel
-/// shifts and [`peel_bit`] masks are defensive, not load-bearing) instead of
-/// on every op.
-#[inline]
-fn wide_add(dst: &mut Wide, src: &Wide) {
-    for (d, s) in dst.iter_mut().zip(src) {
-        *d = d.wrapping_add(*s);
-    }
-}
-
-/// `dst −= src` lane-wise, mod 2^32 (see [`wide_add`] for masking).
-#[inline]
-fn wide_sub(dst: &mut Wide, src: &Wide) {
-    for (d, s) in dst.iter_mut().zip(src) {
-        *d = d.wrapping_sub(*s);
-    }
-}
-
-/// `dst += c·src` lane-wise, mod 2^32 (see [`wide_add`] for masking).
-#[inline]
-fn wide_madd(dst: &mut Wide, c: u32, src: &Wide) {
-    if c == 0 {
-        return;
-    }
-    for (d, s) in dst.iter_mut().zip(src) {
-        *d = d.wrapping_add(c.wrapping_mul(*s));
-    }
-}
-
-/// Pack bit `bit` of every lane into a Z₂ label's words.
-#[inline]
-fn pack_lane_bit(lanes: &Wide, bit: u32) -> Z2 {
-    let mut out = [0u64; 2];
-    for (w, chunk) in lanes.chunks_exact(64).enumerate() {
-        let mut acc = 0u64;
-        for (i, &l) in chunk.iter().enumerate() {
-            acc |= (((l >> bit) & 1) as u64) << i;
-        }
-        out[w] = acc;
-    }
-    out
-}
-
-/// CCRH pad of one width-`l` cast switch, unpacked to lanes: the CF Z_{2^l}
-/// analogue of [`hash::hash_z2`] for wider payloads on bare words. `nonce` is
-/// a solo-domain id (bit 63 clear), globally fresh.
-fn hash_cast(ctrl: &Z2, nonce: u64, l: u32, out: &mut Wide) {
-    debug_assert!(nonce < (1u64 << 63), "solo nonce uses the bulk-domain bit");
-    debug_assert!((2..=32).contains(&l));
-    let mut seed = [0u8; 16];
-    seed[0..8].copy_from_slice(&ctrl[0].to_le_bytes());
-    seed[8..16].copy_from_slice(&ctrl[1].to_le_bytes());
-    let words = (LAMBDA * l as usize).div_ceil(64);
-    // 16 tail bytes keep the NEON gather's last 16-byte load in bounds.
-    let mut buf = [0u8; 512 + 16];
-    expand(seed, nonce, &mut buf[..words * 8]);
-    #[cfg(target_arch = "aarch64")]
-    if neon_unpack_ok(l as usize) {
-        unpack_even_k_neon(&buf, l as usize, out);
-        return;
-    }
-    unpack_generic(&buf[..words * 8], l as usize, out);
-}
-
-#[inline]
-fn lane_mask(k: u32) -> u32 {
-    if k >= 32 { u32::MAX } else { (1u32 << k) - 1 }
-}
-
-/// Scalar unpack of LAMBDA packed `k`-bit coordinates from little-endian
-/// bytes into u32 lanes (branchless two-word windows; cf. label.rs).
-fn unpack_generic(bytes: &[u8], k: usize, dst: &mut Wide) {
-    let mut padded = [0u64; LAMBDA * 32 / 64 + 1];
-    for (w, slot) in padded.iter_mut().enumerate().take(bytes.len() / 8) {
-        *slot = u64::from_le_bytes(bytes[w * 8..(w + 1) * 8].try_into().unwrap());
-    }
-    let mask = lane_mask(k as u32) as u64;
-    for (i, lane) in dst.iter_mut().enumerate() {
-        let bit = i * k;
-        let window = (padded[bit >> 6] as u128) | ((padded[(bit >> 6) + 1] as u128) << 64);
-        *lane = ((window >> (bit & 63)) as u64 & mask) as u32;
-    }
-}
-
-/// True iff [`unpack_even_k_neon`] handles width `k`: even (group bases stay
-/// byte-aligned) and every lane's `shift + k` fits the 32-bit gather window
-/// (this excludes exactly k = 30, whose lane-1 shift of 6 overflows it).
-#[cfg(target_arch = "aarch64")]
-fn neon_unpack_ok(k: usize) -> bool {
-    if !k.is_multiple_of(2) || !(2..=32).contains(&k) {
-        return false;
-    }
-    (0..4).all(|j| ((j * k) & 7) + k <= 32)
-}
-
-/// NEON unpack for even `k` (2 ≤ k ≤ 32): per 4-lane group, one unaligned
-/// 16-byte load at the group's (byte-aligned) base, a TBL byte-gather of each
-/// lane's 4-byte window, per-lane right shifts and a mask.
-///
-/// Even k makes every group base byte-aligned (`4k ≡ 0 (mod 8)`), so the
-/// gather indices and shifts are group-invariant. `scratch` must be at least
-/// `(124·k)/8 + 16` bytes (the 512-byte hash scratch always is for k ≤ 32);
-/// lane 3's window ends at byte 3k/8 + 4 ≤ 16 within each load.
-#[cfg(target_arch = "aarch64")]
-fn unpack_even_k_neon(scratch: &[u8], k: usize, dst: &mut Wide) {
-    use std::arch::aarch64::*;
-    debug_assert!(neon_unpack_ok(k));
-    assert!(
-        scratch.len() >= (124 * k) / 8 + 16,
-        "unpack scratch too short"
-    );
-
-    // Lane j of a group reads bytes (j·k)/8 .. +4, then shifts by (j·k) % 8.
-    let mut idx = [0u8; 16];
-    let mut shifts = [0i32; 4];
-    for j in 0..4 {
-        let bit = j * k;
-        for b in 0..4 {
-            idx[j * 4 + b] = ((bit >> 3) + b) as u8;
-        }
-        shifts[j] = -((bit & 7) as i32);
-    }
-    // SAFETY: NEON is part of the crate's aarch64 baseline (the CCRH core
-    // already requires it). All loads are in-bounds: group g loads 16 bytes
-    // at (g·4·k)/8 with g ≤ 31, bounded by the assert above.
-    unsafe {
-        let idx_v = vld1q_u8(idx.as_ptr());
-        let shift_v = vld1q_s32(shifts.as_ptr());
-        let mask_v = vdupq_n_u32(lane_mask(k as u32));
-        for g in 0..(LAMBDA / 4) {
-            let base = (g * 4 * k) >> 3;
-            let window = vld1q_u8(scratch.as_ptr().add(base));
-            let gathered = vreinterpretq_u32_u8(vqtbl1q_u8(window, idx_v));
-            let aligned = vshlq_u32(gathered, shift_v);
-            let lanes = vandq_u32(aligned, mask_v);
-            vst1q_u32(dst.as_mut_ptr().add(g * 4), lanes);
-        }
-    }
-}
 
 /// Sub-chunk widths for decomposing an `ell`-bit value, each at most
 /// `max_width` bits (greedy: `ell = 22, max = 8 → [8, 8, 6]`).
@@ -221,72 +25,6 @@ pub fn compute_sub_widths(ell: u32, max_width: u32) -> Vec<u32> {
         remaining -= w;
     }
     widths
-}
-
-/// `bit i` of `sub − res` as a Z₂ label: the peel step
-/// `mod2k(div2k(sub − mod2k(res, k), i), 1)` on lanes. `sub` is width-k
-/// lanes, `res` width-l lanes (l ≥ k).
-#[inline]
-fn peel_bit(sub: &Wide, res: &Wide, i: u32, k: u32) -> Z2 {
-    let mask_k = ((1u64 << k) - 1) as u32;
-    let mut out = [0u64; 2];
-    for (w, (sc, rc)) in sub
-        .chunks_exact(64)
-        .zip(res.chunks_exact(64))
-        .enumerate()
-    {
-        let mut acc = 0u64;
-        for (idx, (&s, &r)) in sc.iter().zip(rc).enumerate() {
-            let d = s.wrapping_sub(r) & mask_k;
-            acc |= (((d >> i) & 1) as u64) << idx;
-        }
-        out[w] = acc;
-    }
-    out
-}
-
-// ---- cost / nonce accounting ----
-
-/// Garbled-material + hash footprint of one step: a CF switch on Z_{2^w}
-/// costs `w` CCRH blocks, a CF join `w` λ-bit units of communication.
-#[derive(Clone, Copy, Debug, Default)]
-pub struct StepCost {
-    /// Bits the garbler emits (join diffs).
-    pub program_bits: usize,
-    /// CF join width, in `lg|G|` units.
-    pub join_complexity_cf: usize,
-    /// CCRH blocks charged by the ledger (garbler-side: every switch).
-    pub hash_count_cf: usize,
-}
-
-impl StepCost {
-    fn add(&mut self, o: StepCost) {
-        self.program_bits += o.program_bits;
-        self.join_complexity_cf += o.join_complexity_cf;
-        self.hash_count_cf += o.hash_count_cf;
-    }
-}
-
-/// One tree-plus-casts stage: `n_bits` one-hot bits, casts at width `l`.
-fn stage_cost(n_bits: u32, l: u32) -> StepCost {
-    let n = 1usize << n_bits;
-    StepCost {
-        program_bits: (n_bits as usize - 1 + l as usize) * LAMBDA,
-        join_complexity_cf: n_bits as usize - 1 + l as usize,
-        hash_count_cf: (n - 2) + l as usize * n,
-    }
-}
-
-/// Bulk-domain ids one tree consumes (the garbler hashes every slot of levels
-/// 1..k−1; the evaluator indexes the same window by slot position).
-#[inline]
-fn tree_ids(k: u32) -> u64 {
-    (1u64 << k) - 2
-}
-
-/// (bulk ids, solo ids) consumed by one chunk step over `n_bits` input bits.
-pub fn chunk_nonces(n_bits: u32) -> (u64, u64) {
-    (tree_ids(n_bits), 1u64 << n_bits)
 }
 
 /// The extract step's lane machinery represents width-`l` wires as u32
@@ -311,73 +49,6 @@ pub fn extract_nonces(sub_widths: &[u32]) -> (u64, u64) {
     }
     (bulk, solo)
 }
-
-// ---- the shared tree/cast stage ----
-
-/// Garbler side of one doubling tree: level-major, hashes EVERY slot.
-/// Returns (leaf masks, per-level y-XOR sums `⊕_j y_m[j]` for levels 1..k−1).
-///
-/// `lvl1` is the level-1 pair `[X_{b0} ⊕ Δ₂, X_{b0}]` (the affine base
-/// one-hot). Tree ids are `bulk_base + (2^m − 2) + j` for slot `j` of level
-/// `m` — position-indexed so the evaluator lands on the same pads.
-fn tree_garble(k: u32, lvl1: [Z2; 2], bulk_base: u64) -> (Vec<Z2>, Vec<Z2>) {
-    let mut lvl: Vec<Z2> = lvl1.to_vec();
-    let mut ysums: Vec<Z2> = Vec::with_capacity(k as usize - 1);
-    for m in 1..k {
-        let width = 1usize << m;
-        let base = bulk_base + ((1u64 << m) - 2);
-        let mut next = vec![[0u64; 2]; width * 2];
-        let mut ysum = [0u64; 2];
-        for j in 0..width {
-            let y = hash::hash_z2(&lvl[j], base + j as u64);
-            ysum[0] ^= y[0];
-            ysum[1] ^= y[1];
-            next[j] = [lvl[j][0] ^ y[0], lvl[j][1] ^ y[1]];
-            next[j + width] = y;
-        }
-        ysums.push(ysum);
-        lvl = next;
-    }
-    (lvl, ysums)
-}
-
-/// Width-`l` residue bank over a full cast array (garbler side, which holds
-/// every cast): halving class sums with the incremental weighted chain
-/// `res_{j+1} = res_j + 2^j·H_j`. Returns `(res_1..=res_k, root)`; `res_k` is
-/// the upcast functional `Σ p·casts[p]`, `root = Σ_p casts[p]`. Consumes the
-/// casts (halved in place).
-fn residues_all(mut acc: Vec<Wide>) -> (Vec<Wide>, Wide) {
-    let k = acc.len().trailing_zeros();
-    let mut res_chain: Vec<Wide> = Vec::with_capacity(k as usize);
-    // H_j for j = k−1 down to 1, gathered during halving.
-    let mut highs: Vec<Wide> = Vec::with_capacity(k as usize);
-    while acc.len() > 2 {
-        let mid = acc.len() / 2;
-        let mut high = [0u32; LAMBDA];
-        for d in 0..mid {
-            let (lo, hi) = acc.split_at_mut(mid);
-            wide_add(&mut high, &hi[d]);
-            wide_add(&mut lo[d], &hi[d]);
-        }
-        acc.truncate(mid);
-        highs.push(high);
-    }
-    // Level 1: res_1 = S_1[1]; H_0-style entry not needed (bit 0 is native).
-    let mut res = acc[1];
-    res_chain.push(res);
-    // res_{j+1} = res_j + 2^j·H_j, H_j in reverse gather order.
-    for (j, high) in (1..k).zip(highs.iter().rev()) {
-        let mut r = res;
-        wide_madd(&mut r, 1u32 << j, high);
-        res_chain.push(r);
-        res = r;
-    }
-    let mut root = acc[0];
-    wide_add(&mut root, &acc[1]);
-    (res_chain, root)
-}
-
-// ---- extract step ----
 
 /// Garbler-side output of one extract step.
 #[derive(Debug)]
@@ -775,167 +446,15 @@ fn expand_and_fold_class(
     }
 }
 
-// ---- chunk conversion (bin_to_word) ----
-
-/// Garbler-side output of one chunk-conversion step.
-#[derive(Debug)]
-pub struct ChunkGarbleOutput {
-    /// Mask of the chunk word (CF Z_{2^ell}).
-    pub word_mask: Label,
-    /// The n−1 scale-join diffs and the width-ell pin diff.
-    pub scale: Vec<Z2>,
-    /// Pin-join diff (width ell).
-    pub pin: Wide,
-    /// Ledger footprint.
-    pub cost: StepCost,
-}
-
-/// Garbler side: pack `bits` (CF Z₂ masks, LSB-first) into a Z_{2^ell}
-/// word — the straight-line `bin_to_word` (one-hot tree over the input bits,
-/// width-ell casts, `Σ p·A_p`).
-pub fn chunk_batch_garble(
-    bit_masks: &[Label],
-    ell: u32,
-    delta: u128,
-    nonce_bulk: u64,
-    nonce_solo: u64,
-) -> ChunkGarbleOutput {
-    let nb = bit_masks.len() as u32;
-    assert!(
-        (2..=31).contains(&ell) && ell >= nb,
-        "chunk kernel requires 2 <= nb <= ell <= 31 (got nb={nb}, ell={ell})"
-    );
-    let mask_l = ((1u64 << ell) - 1) as u32;
-    let d2 = {
-        let d = delta_r(delta, 2);
-        let raw = d.raw_bits();
-        [raw[0], raw[1]]
-    };
-    let b0 = z2_of(&bit_masks[0]);
-    let lvl1 = [[b0[0] ^ d2[0], b0[1] ^ d2[1]], b0];
-    let (leaves, ysums) = tree_garble(nb, lvl1, nonce_bulk);
-
-    let n = 1usize << nb;
-    let mut casts = vec![[0u32; LAMBDA]; n];
-    for (pp, cast) in casts.iter_mut().enumerate() {
-        hash_cast(&leaves[pp], nonce_solo + pp as u64, ell, cast);
-    }
-    let (res_chain, root) = residues_all(casts);
-    let word_mask = res_chain[nb as usize - 1]; // Σ p·A_p
-
-    let scale: Vec<Z2> = ysums
-        .iter()
-        .zip(bit_masks.iter().skip(1))
-        .map(|(y, b)| {
-            let bw = z2_of(b);
-            [y[0] ^ bw[0], y[1] ^ bw[1]]
-        })
-        .collect();
-    let mut pin = root;
-    let dl = delta_r(delta, 1u64 << ell);
-    wide_add(&mut pin, &wide_of(&dl));
-    for lane in pin.iter_mut() {
-        *lane &= mask_l;
-    }
-
-    ChunkGarbleOutput {
-        word_mask: wide_label(&word_mask, 1u64 << ell),
-        scale,
-        pin,
-        cost: stage_cost(nb, ell),
-    }
-}
-
-/// Evaluator side: the label-side mirror of [`chunk_batch_garble`].
-/// `value` is the chunk's cleartext value (the evaluator knows its bits).
-pub fn chunk_batch_eval(
-    bit_labels: &[Label],
-    value: u64,
-    ell: u32,
-    g_scale: &[Z2],
-    g_pin: &Wide,
-    nonce_bulk: u64,
-    nonce_solo: u64,
-) -> Label {
-    let nb = bit_labels.len() as u32;
-    assert!(
-        (2..=31).contains(&ell) && ell >= nb,
-        "chunk kernel requires 2 <= nb <= ell <= 31 (got nb={nb}, ell={ell})"
-    );
-
-    // Tree: expand every level; the hot slot solves through the scale join.
-    let b0 = z2_of(&bit_labels[0]);
-    let mut lvl: Vec<Z2> = vec![b0, b0];
-    for m in 1..nb {
-        let width = 1usize << m;
-        let base = nonce_bulk + ((1u64 << m) - 2);
-        let hot = (value & ((1u64 << m) - 1)) as usize;
-        let mut next = vec![[0u64; 2]; width * 2];
-        let mut ysum = [0u64; 2];
-        for (j, bj) in lvl.iter().enumerate() {
-            if j == hot {
-                continue;
-            }
-            let y = hash::hash_z2(bj, base + j as u64);
-            ysum[0] ^= y[0];
-            ysum[1] ^= y[1];
-            next[j] = [bj[0] ^ y[0], bj[1] ^ y[1]];
-            next[j + width] = y;
-        }
-        let bm = z2_of(&bit_labels[m as usize]);
-        let d = g_scale[m as usize - 1];
-        let y_hot = [bm[0] ^ d[0] ^ ysum[0], bm[1] ^ d[1] ^ ysum[1]];
-        next[hot] = [lvl[hot][0] ^ y_hot[0], lvl[hot][1] ^ y_hot[1]];
-        next[hot + width] = y_hot;
-        lvl = next;
-    }
-
-    // Casts of the non-hot leaves; the hot cast solves through the pin join,
-    // and the word label is Σ p·A_p with the hot term substituted.
-    let hot = (value & ((1u64 << nb) - 1)) as usize;
-    let mut word = [0u32; LAMBDA];
-    let mut t_sum = [0u32; LAMBDA];
-    let mut cast = [0u32; LAMBDA];
-    for (pp, leaf) in lvl.iter().enumerate() {
-        if pp == hot {
-            continue;
-        }
-        hash_cast(leaf, nonce_solo + pp as u64, ell, &mut cast);
-        wide_madd(&mut word, pp as u32, &cast);
-        wide_add(&mut t_sum, &cast);
-    }
-    let mut hot_cast = *g_pin; // L_root = pin diff (constant-1 label is 0)
-    wide_sub(&mut hot_cast, &t_sum);
-    wide_madd(&mut word, hot as u32, &hot_cast);
-
-    wide_label(&word, 1u64 << ell)
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
     use crate::label::{self, LAMBDA};
     use rand::Rng;
 
-    /// The NEON even-k unpacker must agree with the scalar window loop on
-    /// every width it claims (the cast pads flow through this).
-    #[cfg(target_arch = "aarch64")]
-    #[test]
-    fn test_unpack_even_k_neon_matches_generic() {
-        let mut rng = rand::rng();
-        let mut scratch = vec![0u8; 512];
-        for round in 0..8 {
-            rng.fill(&mut scratch[..]);
-            for k in (2..=32).step_by(2).filter(|&k| neon_unpack_ok(k)) {
-                let bytes = (LAMBDA * k).div_ceil(64) * 8;
-                let mut want = [0u32; LAMBDA];
-                unpack_generic(&scratch[..bytes], k, &mut want);
-                let mut got = [0u32; LAMBDA];
-                unpack_even_k_neon(&scratch, k, &mut got);
-                assert_eq!(want, got, "k={k}, round={round}");
-            }
-            assert!(!neon_unpack_ok(30) && !neon_unpack_ok(7));
-        }
+    fn rand_wide(rng: &mut impl Rng, modulus: u64) -> Label {
+        let coords: Vec<u64> = (0..LAMBDA).map(|_| rng.random_range(0..modulus)).collect();
+        Label::from_coords(&coords, modulus)
     }
 
     #[test]
@@ -944,64 +463,6 @@ mod tests {
         assert_eq!(compute_sub_widths(16, 8), vec![8, 8]);
         assert_eq!(compute_sub_widths(3, 8), vec![3]);
         assert_eq!(compute_sub_widths(4, 1), vec![1, 1, 1, 1]);
-    }
-
-    fn rand_z2(rng: &mut impl Rng) -> Label {
-        let coords: Vec<u64> = (0..LAMBDA).map(|_| rng.random_range(0..2u64)).collect();
-        Label::from_coords(&coords, 2)
-    }
-
-    fn rand_wide(rng: &mut impl Rng, modulus: u64) -> Label {
-        let coords: Vec<u64> = (0..LAMBDA)
-            .map(|_| rng.random_range(0..modulus))
-            .collect();
-        Label::from_coords(&coords, modulus)
-    }
-
-    /// Chunk step: word label = word mask + v·Δ for every value, at the
-    /// production shape (nb = 8, ell = 22) and a small one.
-    #[test]
-    fn test_chunk_label_mask_invariant_production() {
-        chunk_invariant_shape(8, 22);
-    }
-
-    #[test]
-    fn test_chunk_label_mask_invariant() {
-        chunk_invariant_shape(4, 12);
-    }
-
-    fn chunk_invariant_shape(nb: u32, ell: u32) {
-        let mut rng = rand::rng();
-        let delta: u128 = rng.random::<u128>() | 1;
-        let d2 = delta_r(delta, 2);
-        let dl = delta_r(delta, 1u64 << ell);
-
-        let bit_masks: Vec<Label> = (0..nb).map(|_| rand_z2(&mut rng)).collect();
-        let g = chunk_batch_garble(&bit_masks, ell, delta, 5_000, 9_000);
-        let lim = 1u32 << ell;
-        assert!(g.pin.iter().all(|&lane| lane < lim), "pin lane not canonical");
-        assert_eq!(g.cost.join_complexity_cf as u32, nb - 1 + ell);
-        assert_eq!(
-            g.cost.hash_count_cf as u64,
-            (1u64 << nb) - 2 + ell as u64 * (1u64 << nb)
-        );
-
-        for v in 0..(1u64 << nb) {
-            let bit_labels: Vec<Label> = bit_masks
-                .iter()
-                .enumerate()
-                .map(|(j, m)| {
-                    if (v >> j) & 1 == 1 {
-                        label::add(m, &d2)
-                    } else {
-                        m.clone()
-                    }
-                })
-                .collect();
-            let w = chunk_batch_eval(&bit_labels, v, ell, &g.scale, &g.pin, 5_000, 9_000);
-            let expect = label::add(&g.word_mask, &label::scalar_mul(v, &dl));
-            assert_eq!(w, expect, "chunk word label ≠ mask + v·Δ for v={v} nb={nb} ell={ell}");
-        }
     }
 
     /// The extract step rejects width-1 sub-chunks loudly (the lane
@@ -1022,7 +483,7 @@ mod tests {
         let mut rng = rand::rng();
         let delta: u128 = rng.random::<u128>() | 1;
         let d2 = delta_r(delta, 2);
-
+    
         for (ell, sub_widths, coeffs) in [
             (8u32, vec![4u32, 4], vec![1u64, 16]),
             (22, vec![8, 8, 6], vec![1, 37, 4123]),
@@ -1031,7 +492,7 @@ mod tests {
             let word_masks: Vec<Label> =
                 (0..coeffs.len()).map(|_| rand_wide(&mut rng, m_ell)).collect();
             let (bulk_n, solo_n) = extract_nonces(&sub_widths);
-
+    
             let g = extract_batch_garble(&word_masks, &coeffs, &sub_widths, delta, 100, 200);
             // Emitted material must be canonical: every pin lane < 2^l, where
             // l is the stage's cast width (the remainder width, or k on the
@@ -1057,7 +518,7 @@ mod tests {
             };
             assert_eq!(g.cost.hash_count_cf, expect_cost);
             let _ = (bulk_n, solo_n);
-
+    
             // Random chunk-word values; r = Σ coeff·v mod 2^ell.
             for _ in 0..24 {
                 let values: Vec<u64> =
@@ -1072,7 +533,7 @@ mod tests {
                     .zip(&values)
                     .map(|(m, &v)| label::add(m, &label::scalar_mul(v, &dl_ell)))
                     .collect();
-
+    
                 let (fbh, fbits) = extract_batch_eval(
                     &word_labels,
                     &coeffs,
@@ -1082,7 +543,7 @@ mod tests {
                     100,
                     200,
                 );
-
+    
                 // first_bin_hot: label = mask + [p == r mod 2^{w0}]·Δ₂.
                 let w0 = sub_widths[0];
                 let low = (r & ((1u64 << w0) - 1)) as usize;
