+++ src/gc/fold.rs
